+++ src/app/components/configuration/account/account-config/account-config.component.spec.ts
@@ -2,7 +2,7 @@ import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { fakeAsync, getTestBed, tick } from '@angular/core/testing';
 import { AccountConfigComponent } from './account-config.component';
 import { Account} from 'src/app/shared/types/account';
-import { FluxStore } from 'src/app/shared/services/flux-store';
+import { FluxStore } from 'src/app/model/flux-store';
 import { CommonModule } from '@angular/common';
 import { SharedModule } from 'src/app/components/shared_components/shared.module';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
