+++ src/app/components/configuration/account/account-config/account-config.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnDestroy, OnInit, ViewChild } from '@angular/core';
 import { Subscription } from 'rxjs';
-import { FluxStore } from 'src/app/shared/services/flux-store';
+import { FluxStore } from 'src/app/model/flux-store';
 import { Account } from 'src/app/shared/types/account';
 import { csvMask } from 'src/app/shared/types/csvMask';
 import { AccountFormComponent } from '../account-form/account-form.component';
