+++ src/app/components/configuration/account/account-form/account-form.component.html
@@ -47,7 +47,7 @@
                 required>
             </div>
             <div *ngIf="(shortName.invalid && shortName.touched) || shortName.dirty">
-              <p *ngIf="shortName.invalid" class="validation">A maximum short length of 6 characters</p>
+              <p *ngIf="shortName.invalid" class="validation">A minumum 1 characters and maximum 8 characters</p>
             </div>
           </div>
           <div class="column">
@@ -103,7 +103,7 @@
             </p>
           </div>
           <div *ngIf="(initialValue.invalid && initialValue.touched) || initialValue.dirty">
-            <p *ngIf="initialValue.invalid" class="validation">Must be a number</p>
+            <p *ngIf="initialValue.invalid" class="validation">Is required and must be a number</p>
           </div>
         </div>
         <div class="field">
@@ -122,7 +122,6 @@
             </select>
           </div>
         </div>
-
       </div>
     </form>
