+++ src/app/components/configuration/account/account-form/account-form.component.ts
@@ -2,7 +2,7 @@ import { Component, ElementRef, Inject, Input, OnChanges, OnInit, SimpleChanges,
 import { FormControl, FormGroup, FormGroupDirective, Validators } from '@angular/forms';
 import { Subject, Subscription } from 'rxjs';
 import { fluxDispatcherToken } from 'src/app/shared/helpers/flux.configuration';
-import { FluxStore } from 'src/app/shared/services/flux-store';
+import { FluxStore } from 'src/app/model/flux-store';
 import { Account, AccountColors } from 'src/app/shared/types/account';
 import { csvMask } from 'src/app/shared/types/csvMask';
 import { FluxAction, FluxActionTypes } from 'src/app/shared/types/actions.type';
