+++ src/app/components/configuration/csv/csv-config/csv-config.component.ts
@@ -1,6 +1,6 @@
 import { Component, Input, OnDestroy, OnInit, ViewChild } from '@angular/core';
 import { Subscription } from 'rxjs';
-import { FluxStore } from 'src/app/shared/services/flux-store';
+import { FluxStore } from 'src/app/model/flux-store';
 import { csvMask } from 'src/app/shared/types/csvMask';
 import { CsvFormComponent } from '../csv-form/csv-form.component';
