+++ src/app/components/dashboard/account/account.component.spec.ts
@@ -2,7 +2,7 @@ import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { initializeApp, provideFirebaseApp } from '@angular/fire/app';
 import { getFirestore, provideFirestore } from '@angular/fire/firestore';
 import { FLUX_CONFIG } from 'src/app/shared/helpers/flux.configuration';
-import { FluxStore } from 'src/app/shared/services/flux-store';
+import { FluxStore } from 'src/app/model/flux-store';
 import { environment } from 'src/environments/environment';
 import { fluxDispatcherToken } from 'src/app/shared/helpers/flux.configuration';
 import { Subject, Subscription } from 'rxjs';
