+++ src/app/components/dashboard/transaction-list/transaction-form/transaction-form.component.html
@@ -27,6 +27,8 @@
                 required>
             </div>
           </div>
+
+
           <div class="column">
             <label class="label">Amount</label>
             <div class="field has-addons">
