+++ src/app/components/dashboard/transaction-list/transaction-list.component.scss
@@ -10,7 +10,3 @@
 .is-transaction{
   cursor: pointer;
 }
-
-.is-transaction:hover{
-  background-color: bisque;
-}
