+++ src/app/shared/services/utility.service.ts
@@ -31,4 +31,5 @@ export class UtilityService {
     return Number(currentValue).toFixed(2)
   }
 
+
 }
