+++ src/styles/general.scss
@@ -54,7 +54,7 @@ table.transaction-table {
 }
 
 .validation{
-  color : red;
+  color : red !important;
   font-weight: 600;
   font-size: 0.875em;
 }
